requirements.txt
ansible==2.7.10
cryptography==2.6.1
GitPython==2.1.11
